feat back: add API REST crud in entity Segmento

+++ Clininca_veterinaria_2023_DDD/Domain/Interfaces/Generics/InterfaceGeneric.cs
@@ -14,7 +14,7 @@ namespace Domain.Interfaces.Generics
 
         Task Delete(T objeto);
 
-        Task<T> GetEntityById(int id);
+        Task<T> GetEntityById(long id);
 
         Task<List<T>> List(); 
     }
